% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyBars.R
\name{anyBars}
\alias{anyBars}
\title{An internal function to detect the random effects component from an object of class formula}
\usage{
anyBars(term)
}
\arguments{
\item{term}{formula to be processed}
}
\description{
An internal function to detect the random effects component from an object of class formula
}
\examples{
#Internal function only
#no examples
#
#
}
\keyword{effects}
\keyword{models}
\keyword{random}
