% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_gamma_hurdle_e_ind.R
\name{beta_gamma_hurdle_e_ind}
\alias{beta_gamma_hurdle_e_ind}
\title{An internal function to write BUGS code for bivariate beta-gamma with an hurdle model for the effects (independence)}
\usage{
beta_gamma_hurdle_e_ind(type)
}
\arguments{
\item{type}{Type of structural value mechanism assumed. Choices are Structural Completely At Random (SCAR),
and Structural At Random (MNAR)}
}
\description{
This function writes in the current WD a txt file 
for the bivariate independent beta-gamma with an hurdle model for the effects.
}
\examples{
#Internal function only
#No examples
#
#
}
\keyword{Hurdle}
\keyword{JAGS}
\keyword{models}
