% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_gamma_ind.R
\name{beta_gamma_ind}
\alias{beta_gamma_ind}
\title{An internal function to write BUGS code for beta-gamma model with natural-scaled variables (independence)}
\usage{
beta_gamma_ind(type)
}
\arguments{
\item{type}{Type of missingness mechanism assumed. Choices are Missing Completely At Random (MCAR),
Missing At Random (MAR), Missing Not At Random (MNAR). For 'MNAR' alternative versions are available 
depending on whether the mechanism for only one variable is condiered, that is for the effects (MNAR_eff)
or the costs (MNAR_cost), or also covariates are included either for the effects (MNAR_eff_cov),
the costs (MNAR_cost), or both (MNAR_cov)}
}
\description{
This function writes in the current WD a txt file 
for the bivariate independent beta-gamma model with natural-scaled outcome variables.
}
\examples{
#Internal function only
#No examples
#
#
}
\keyword{BUGS}
\keyword{JAGS}
\keyword{models}
