% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_gamma_selection_ind.R
\name{beta_gamma_selection_ind}
\alias{beta_gamma_selection_ind}
\title{An internal function to write BUGS code for bivariate beta-gamma selection model (independence)}
\usage{
beta_gamma_selection_ind(type)
}
\arguments{
\item{type}{Type of missingness mechanism assumed. Choices are Missing At Random (MAR), Missing Not At Random for the effects (MNAR_eff),
Missing Not At Random for the costs (MNAR_cost), and Missing Not At Random for both (MNAR)}
}
\description{
This function writes in the current WD a txt file 
for the bivariate independent beta-gamma selection model for both effects and costs.
}
\examples{
#Internal function only
#No examples
#
#
}
\keyword{JAGS}
\keyword{Selection}
\keyword{models}
