% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_normal_hurdle_c_ind.R
\name{beta_normal_hurdle_c_ind}
\alias{beta_normal_hurdle_c_ind}
\title{An internal function to write BUGS code for bivariate beta-normal with an hurdle model for the costs (independence)}
\usage{
beta_normal_hurdle_c_ind(type)
}
\arguments{
\item{type}{Type of structural value mechanism assumed. Choices are Structural Completely At Random (SCAR),
and Structural At Random (MNAR)}
}
\description{
This function writes in the current WD a txt file 
for the independent joint beta-normal with an hurdle model for the costs.
}
\examples{
#Internal function only
#No examples
#
#
}
\keyword{Hurdle}
\keyword{JAGS}
\keyword{models}
