% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_normal_hurdle_ec_joint.R
\name{beta_normal_hurdle_ec_joint}
\alias{beta_normal_hurdle_ec_joint}
\title{An internal function to write BUGS code for bivariate beta-normal hurdle model (joint)}
\usage{
beta_normal_hurdle_ec_joint(type)
}
\arguments{
\item{type}{Type of structural value mechanism assumed. Choices are Structural Completely At Random (SCAR),
and Structural At Random (MNAR)}
}
\description{
This function writes in the current WD a txt file 
for the bivariate joint beta-normal hurdle model for both effects and costs.
}
\examples{
#Internal function only
#No examples
#
#
}
\keyword{Hurdle}
\keyword{JAGS}
\keyword{models}
