% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_checks.R
\name{diagnostic_checks}
\alias{diagnostic_checks}
\title{Diagnostic checks for assessing MCMC convergence}
\usage{
diagnostic_checks(x, type = "histogram", param = "all", theme = NULL, ...)
}
\arguments{
\item{x}{An object of class "missingHE" containing the posterior results of a full Bayesian model implemented usgin the function \code{\link{run_model}}}

\item{type}{type of diagnostic check to be plotted for model parameters, mostly taken from the packages \strong{ggmcmc} and \strong{mcmcplots}. Available choices include: 'histogram' for histrogram plots,
'denplot' for density plots, 'traceplot' for trace plots and 'acf' for autocorrelation plots, 'running' for running mean plots,
'compare' for comparing the distribution of the whole chain with only its last part, 'cross' for crosscorrelation plots, 'Rhat' for the potential scale reduction factor, 'geweke' for the geweke diagnostic,
'pairs' for posterior correlation among the parameters,'caterpillar' for caterpillar plots. In addition the class 'summary' provides an overview of some of the most popular
diagnostic checks for each parameter selected.}

\item{param}{Name of the family of parameters to process, as given by a regular expression. For example the mean parameters 
for the effect and cost variables can be specified using 'mu.e' and 'mu.c', respectively. Different types of
models may have different parameters depending on the assumed distributions and missing data mechanisms. 
To see a complete list of all possible paraemters by types of models assumed see details.}

\item{theme}{Type of ggplot theme among some pre-defined themes, mostly taken from the package \strong{ggthemes}. For a full list of available themes see details.}

\item{...}{additional parameters that can be provided to manage the output of \code{diagnostic_checks}. 
These are mostly taken from the package \strong{ggmcmc} and are for graphical purposes.}
}
\value{
A \code{ggplot} object containing the plots specified in the argument \code{class}
}
\description{
Convergence issues need to be assessed every time a  probabilistic model is implemented (via MCMC). The focus 
is restricted to full Bayesian models in cost-effectiveness analyses run using the function \code{\link{run_model}}, 
with convergence of the chains that is assessed through Visual checks of the posterior distribution of the parameters of interest,
such as density plots, trace plots, autocrrelation plots, etc. Other types of posterior checks are related to some summary MCMC statistics 
able to detect possible issues in the convergence of the algorithm, such as the potential scale reduction factor or the effective sample size.
\code{diagnostic_checks} uses many different types of diagnostic tools and statistics to assess model convergence from MCMC algorithms
using functions contained in the package \code{ggmcmc} and \code{mcmcplots}. Graphics display is managed and tailored using functions contained 
in the package \code{ggplot2} and \code{ggthemes}.
}
\details{
Depending on the types of plots specified in the argument \code{class} the output of \code{diagnostic_checks} can produce
different combinations of MCMC visual posterior checks for the parameters of interest indicated in the argument \code{param}.
For a full list of the available plots see the description of the argument \code{class} or see the corresponding plots in the package \strong{ggmcmc}.

The parameters that can be assessed through \code{diagnostic_checks} are only those inlcuded in the object \code{x} (see Arguments)
and are related to different model aspects in the setting of cost-effectiveness analyses. Specifically, in order to assess each parameter,
a different character name must be specified in the argument \code{param}. The available names and the parameters associated with them are:
\itemize{
\item "mu.e" the mean parameters of the effect variables in the two treatment arms.
\item "mu.c" the mean parameters of the cost variables in the two treatment arms.
\item "sd.e" the standard deviation parameters of the effect variables in the two treatment arms.
\item "sd.c" the standard deviation parameters of the cost variables in the two treatment arms.
\item "beta0.e" the marginal mean parameters for the effect variables in the two treatment arms (only when covariates were included in the model).
\item "beta.e" the covariate coefficient parameters for the effect variables in the two treatment arms (only when covariates were included in the model).
\item "beta0.c" the marginal mean parameters for the cost variables in the two treatment arms (only when covariates were included in the model).
\item "beta.c" the covariate coefficient parameters for the cost variables in the two treatment arms (only when covariates were included in the model).
\item "gamma0.e" the baseline parameters of the missingness mechanism for the effect variables in the two treatment arms.
\item "gamma0.c" the baseline parameters of the missingness mechanism for the cost variables in the two treatment arms.
\item "gamma.e" the covariate coefficient parameters of the missingness mechanism for the effect variables in the two treatment arms.
\item "gamma.c" the covariate coefficient parameters of the missingness mechanism for the cost variables in the two treatment arms.
\item "delta.e" the mnar parameters of the missingness mechanism for the effect variables in the two treatment arms.
\item "delta.c" the mnar parameters of the missingness mechanism for the cost variables in the two treatment arms.
\item "all" all available parameters stored in the object \code{x}.
}
To notice that the marginal mean and the covariate coefficient parameters can be accessed only if covariate data were included in the model
and stored in the object \code{x} of class \strong{MissingHE}. Specifically, the mean parameters "beta0.e" and/or "beta0.c" are mere substitute of the 
effect and cost mean parameters in the case covariates are included in the model (must be used as substitutes of "mu.e" and/or "mu.c").
Finally, the parameters associated with the missingness model can be accessed only if an \emph{informative} or MNAR missing data mechanism was assumed
when runinng the model using \code{\link{run_model}}.

The argument \code{theme} allows to customise the graphical aspect of the plots generated by \code{diagnostic_checks} and
allows to choose among a set of possible pre-defined themes taken form the package \strong{ggtheme}. Those available can
be indicated using the following character names: "base","calc","economist","excel","few","538","gdocs","hc","par","pander","solarized","stata","tufte","wsj".
}
\examples{
#For examples see the function run_model
#
#
}
\references{
Gelman, A. Carlin, JB., Stern, HS. Rubin, DB.(2003). \emph{Bayesian Data Analysis, 2nd edition}, CRC Press.

Brooks, S. Gelman, A. Jones, JL. Meng, XL. (2011). \emph{Handbook of Markov Chain Monte Carlo}, CRC/Chapman and Hall.
}
\seealso{
\code{\link[ggmcmc]{ggs}} \code{\link{run_model}}
}
\author{
Andrea Gabrio
}
\keyword{MCMC}
\keyword{checks}
\keyword{convergence}
\keyword{diagnostics}
