% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb.R
\name{fb}
\alias{fb}
\title{An internal function to extract the random effects component from an object of class formula}
\usage{
fb(term)
}
\arguments{
\item{term}{formula to be processed}
}
\description{
An internal function to extract the random effects component from an object of class formula
}
\examples{
#Internal function only
#no examples
#
#
}
\keyword{effects}
\keyword{models}
\keyword{random}
