% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nobars_.R
\name{nobars_}
\alias{nobars_}
\title{An internal function to separate the fixed and random effects components from an object of class formula}
\usage{
nobars_(term)
}
\arguments{
\item{term}{formula to be processed}
}
\description{
An internal function to separate the fixed and random effects components from an object of class formula
}
\examples{
#Internal function only
#no examples
#
#
}
\keyword{effects}
\keyword{models}
\keyword{random}
