% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_gamma_hurdle_c_joint.R
\name{normal_gamma_hurdle_c_joint}
\alias{normal_gamma_hurdle_c_joint}
\title{An internal function to write BUGS code for bivariate normal-gamma with an hurdle model for the costs (joint)}
\usage{
normal_gamma_hurdle_c_joint(type)
}
\arguments{
\item{type}{Type of structural value mechanism assumed. Choices are Structural Completely At Random (SCAR),
and Structural At Random (MNAR)}
}
\description{
This function writes in the current WD a txt file 
for the bivariate joint normal-gamma with an hurdle model for the costs
}
\examples{
#Internal function only
#No examples
#
#
}
\keyword{Hurdle}
\keyword{JAGS}
\keyword{models}
