% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_gamma_hurdle_ec_joint.R
\name{normal_gamma_hurdle_ec_joint}
\alias{normal_gamma_hurdle_ec_joint}
\title{An internal function to write BUGS code for bivariate normal-gamma hurdle model (joint)}
\usage{
normal_gamma_hurdle_ec_joint(type)
}
\arguments{
\item{type}{Type of structural value mechanism assumed. Choices are Structural Completely At Random (SCAR),
and Structural At Random (MNAR). It is also possible to explore some Structural Not At Random (SNAR) assumptions
by providing a structural indicator variable for the missing cases for both effects and costs}
}
\description{
This function writes in the current WD a txt file 
for the bivariate joint normal-gamma hurdle model for both effects and costs.
}
\examples{
#Internal function only
#No examples
#
#
}
\keyword{Hurdle}
\keyword{JAGS}
\keyword{models}
