% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_gamma_selection_joint.R
\name{normal_gamma_selection_joint}
\alias{normal_gamma_selection_joint}
\title{An internal function to write BUGS code for bivariate normal-gamma selection model (joint)}
\usage{
normal_gamma_selection_joint(type)
}
\arguments{
\item{type}{Type of missingness mechanism assumed. Choices are Missing At Random (MAR), Missing Not At Random for the effects (MNAR_eff),
Missing Not At Random for the costs (MNAR_cost), and Missing Not At Random for both (MNAR)}
}
\description{
This function writes in the current WD a txt file 
for the bivariate joint normal-gamma selection model for both effects and costs.
}
\examples{
#Internal function only
#No examples
#
#
}
\keyword{JAGS}
\keyword{Selection}
\keyword{models}
