% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_hurdle_e_joint.R
\name{normal_hurdle_e_joint}
\alias{normal_hurdle_e_joint}
\title{An internal function to write BUGS code for bivariate normal with an hurdle model for the effects (joint)}
\usage{
normal_hurdle_e_joint(type)
}
\arguments{
\item{type}{Type of structural value mechanism assumed. Choices are Structural Completely At Random (SCAR),
and Structural At Random (MNAR)}
}
\description{
This function writes in the current WD a txt file 
for the bivariate joint normal with an hurdle model for the effects
}
\examples{
#Internal function only
#No examples
#
#
}
\keyword{Hurdle}
\keyword{JAGS}
\keyword{models}
