% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pic.R
\name{pic}
\alias{pic}
\title{Predictive information criteria for Bayesian models fitted in \code{JAGS} using the funciton \code{\link{selection}} or \code{\link{hurdle}}}
\usage{
pic(x, criterion = "dic", module = "total")
}
\arguments{
\item{x}{A \code{missingHE} object containing the results of the Bayesian model run using the function \code{\link{selection}} or \code{\link{hurdle}}.}

\item{criterion}{type of information criteria to be produced. Available choices are \code{'dic'} for the Deviance Information Criterion, 
\code{'waic'} for the Widely Applicable Information Criterion, and \code{'looic'} for the Leave-One-Out Information Criterion.}

\item{module}{The modules with respect to which the information criteria should be computed. Available choices are \code{'total'} for the whole model, 
\code{'e'} for the effectiveness variables only, \code{'c'} for the cost variables only, and \code{'both'} for both outcome variables only.}
}
\description{
Efficient approximate leave-one-out cross validation (LOO), deviance information criterion (DIC) and widely applicable criterion (WAIC) for Bayesian models, 
calculated on the observed data.
}
\examples{
 
#For examples see the function selection or hurdle 
# 
# 
}
\author{
Andrea Gabrio
}
\keyword{JAGS}
\keyword{dic}
\keyword{loo}
\keyword{waic}
