% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.missingHE.R
\name{plot.missingHE}
\alias{plot.missingHE}
\title{Plot method for the imputed data contained in the objects of class \code{missingHE}}
\usage{
\method{plot}{missingHE}(
  x,
  prob = c(0.025, 0.975),
  class = "scatter",
  outcome = "both",
  trt = "all",
  time_plot = NULL,
  theme = NULL,
  only.plot = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{missingHE} object containing the results of the Bayesian model for cost-effectiveness analysis.}

\item{prob}{A numeric vector of probabilities representing the upper and lower CI sample quantiles to be calculated and returned for the imputed values.}

\item{class}{Type of the plot comparing the observed and imputed outcome data. Available choices are
'scatter' (default), 'histogram' and 'boxplot' for a scatter, histogram or a boxplot of the observed and average imputed outcome data, respectively.}

\item{outcome}{The outcome variables that should be displayed. Options are: 'both' (default) for both effects and costs; 'effects' or 'costs' for the effects
or costs separately.}

\item{trt}{treatment group for which plots should be displayed. Choices include: 'all' (default) for all groups; 'none' for results across all groups; any 
character or numeric value denoting the treatment group name or index associated with the treatment variable in the original data set.}

\item{time_plot}{Time point for which the graphs should be displayed.}

\item{theme}{Type of ggplot theme among some pre-defined themes, mostly taken from the package \strong{ggthemes}. For a full list of available themes see details.}

\item{only.plot}{Logical. If \code{only.plot} is \code{TRUE} (default) only the plot is returned. When \code{only.plot} is \code{FALSE}, a list containing the plot object and the data is returned instead.}

\item{...}{Additional parameters that can be provided to manage the output of \code{plot.missingHE}.}
}
\value{
A \code{ggplot} object containing the plots specified in the argument \code{class}.
}
\description{
Produces a plot of the observed and imputed values (with credible intervals) for the effect and cost outcomes from a
Bayesian cost-effectiveness analysis model with two treatment arms, implemented using the function \code{\link{selection}}, \code{\link{pattern}}, 
\code{\link{hurdle}} or \code{\link{lmdm}}.
The graphical layout is obtained from the functions contained in the package \strong{ggplot2} and \strong{ggthemes}.
}
\details{
The function produces a plot of the observed and imputed effect and cost data in a two-arm based 
cost-effectiveness model implemented using the function \code{\link{selection}}, \code{\link{pattern}}, \code{\link{hurdle}} or or \code{\link{lmdm}}. 
The purpose of this graph is to visually compare the outcome values for the fully-observed individuals with those imputed by the model for the missing individuals.
For scatter plots, average imputed values are also associated with credible intervals specified in the argument \code{prob}.
The argument \code{theme} allows to customise the graphical aspect of the plots generated by \code{plot.missingHE} and
allows to choose among a set of possible pre-defined themes taken form the package \strong{ggtheme}. For a complete list of the available character names
for each theme and scheme set, see \strong{ggthemes} and \strong{bayesplot}.
}
\examples{
# For examples see the function \code{\link{selection}}, \code{\link{pattern}},
# \code{\link{hurdle}} or \code{\link{lmdm}}
#
#
}
\references{
Daniels, MJ. Hogan, JW. (2008) \emph{Missing Data in Longitudinal Studies: strategies for Bayesian modelling and sensitivity analysis}, CRC/Chapman Hall.

Molenberghs, G. Fitzmaurice, G. Kenward, MG. Tsiatis, A. Verbeke, G. (2015) \emph{Handbook of Missing Data Methodology}, CRC/Chapman Hall.
}
\seealso{
\code{\link{selection}} \code{\link{pattern}} \code{\link{hurdle}} \code{\link{lmdm}} \code{\link{diagnostic}}
}
\author{
Andrea Gabrio
}
\keyword{data}
\keyword{missing}
\keyword{plot}
