% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.missingHE.R
\name{print.missingHE}
\alias{print.missingHE}
\title{Print method for the posterior results contained in the objects of class \code{missingHE}}
\usage{
\method{print}{missingHE}(x, display = "fixed", digits = 3, ...)
}
\arguments{
\item{x}{A \code{missingHE} object containing the results of the Bayesian model run using the function \code{\link{selection}}, \code{\link{pattern}},
\code{\link{hurdle}} or \code{\link{lmdm}}.}

\item{display}{Set of parameters for which posterior summaries are displayed. Choices are: fixed effects ('fixed'), random effects ('random'),
individual log likelihood values ('loglik'), individual conditional parameter values ('conditional') and individual imputed values ('mis').}

\item{digits}{Integer indicating the number of decimal places to be used for rounding (default = 3).}

\item{...}{additional arguments affecting the output produced. For example: \code{param = } JAGS name assigned to the specific model parameters 
to be shown in the printed table.}
}
\description{
Prints the summary table for the model fitted, with the estimate of the parameters and/or missing values.
}
\examples{
 
# For examples see the function \code{\link{selection}}, \code{\link{pattern}}, 
# \code{\link{hurdle}} or \code{\link{lmdm}}
# 
# 
}
\author{
Andrea Gabrio
}
\keyword{JAGS}
\keyword{data}
\keyword{missing}
\keyword{print}
