% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_change.R
\name{prior_change}
\alias{prior_change}
\title{An internal function to change the hyperprior parameters in the model provided by the user depending on the type of
missing data mechanism and outcome distributions assumed}
\usage{
prior_change(type, dist_e, dist_c)
}
\arguments{
\item{type}{type of missingness mechanism assumed. Choices are Missing Completely At Random (MCAR),
Missing At Random (MAR), Missing Not At Random (MNAR). For 'MNAR' alternative versions are available 
depending on whether the mechanism for only one variable is condiered, that is for the effects (MNAR_eff)
or the costs (MNAR_cost), or also covariates are included either for the effects (MNAR_eff_cov),
the costs (MNAR_cost), or both (MNAR_cov). For a complete list of all available hyper parameters 
and types of models see the manual.}

\item{dist_e}{assumed distribution for the effects in the MoA}

\item{dist_c}{assumed distribution for the costs in the MoA}
}
\description{
This function modifies default hyper prior parameter values in the type of model selected according 
to the type of missingness mechanism and distributions for the outcome variables assumed.
}
\examples{
#Internal function only
#no examples
#
#
}
\keyword{distributions}
\keyword{prior}
