% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_hurdle.R
\name{prior_hurdle}
\alias{prior_hurdle}
\title{An internal function to change the hyperprior parameters in the hurdle model provided by the user depending on the type of
structural value mechanism and outcome distributions assumed}
\usage{
prior_hurdle(type, dist_e, dist_c, model_txt_info, model_string_jags)
}
\arguments{
\item{type}{Type of structural value mechanism assumed. Choices are Structural Completely At Random (SCAR) and Structural At Random (SAR). 
For a complete list of all available hyper parameters and types of models see the manual.}

\item{dist_e}{distribution assumed for the effects. Current available choices are: Normal ('norm'), Beta ('beta'), Gamma ('gamma'), Exponential ('exp'),
Weibull ('weibull'), Logistic ('logis'), Poisson ('pois'), Negative Binomial ('nbinom') or Bernoulli ('bern')}

\item{dist_c}{Distribution assumed for the costs. Current available choices are: Normal ('norm'), Gamma ('gamma') or LogNormal ('lnorm')}

\item{model_txt_info}{list containing model specification information used to write the txt file of the JAGS model.}

\item{model_string_jags}{text file of the model.}
}
\description{
This function modifies default hyper prior parameter values in the type of hurdle model selected according 
to the type of structural value mechanism and distributions for the outcomes assumed.
}
\examples{
#Internal function only
#no examples
#
#
}
\keyword{Hurdle}
\keyword{distributions}
\keyword{models}
\keyword{priors}
