% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_long_miss.R
\name{prior_long_miss}
\alias{prior_long_miss}
\title{An internal function to change the hyperprior parameters in the selection model provided by the user depending on the type of
missingness mechanism and outcome distributions assumed}
\usage{
prior_long_miss(
  type,
  dist_e,
  dist_c,
  pe_fixed,
  pc_fixed,
  ze_fixed,
  zc_fixed,
  model_e_random,
  model_c_random,
  model_me_random,
  model_mc_random,
  pe_random,
  pc_random,
  ze_random,
  zc_random
)
}
\arguments{
\item{type}{Type of missingness mechanism assumed. Choices are Missing At Random (MAR), Missing Not At Random for the effects (MNAR_eff),
Missing Not At Random for the costs (MNAR_cost), and Missing Not At Random for both (MNAR). For a complete list of all available hyper parameters 
and types of models see the manual.}

\item{dist_e}{distribution assumed for the effects. Current available choices are: Normal ('norm'), Beta ('beta'), Gamma ('gamma'), Exponential ('exp'),
Weibull ('weibull'), Logistic ('logis'), Poisson ('pois'), Negative Binomial ('nbinom') or Bernoulli ('bern')}

\item{dist_c}{Distribution assumed for the costs. Current available choices are: Normal ('norm'), Gamma ('gamma') or LogNormal ('lnorm')}

\item{pe_fixed}{Number of fixed effects for the effectiveness model}

\item{pc_fixed}{Number of fixed effects for the cost model}

\item{ze_fixed}{Number of fixed effects or the missingness indicators model for the effectiveness}

\item{zc_fixed}{Number of fixed effects or the missingness indicators model for the costs}

\item{model_e_random}{Random effects formula for the effectiveness model}

\item{model_c_random}{Random effects formula for the costs model}

\item{model_me_random}{Random effects formula for the missingness indicators model for the effectiveness}

\item{model_mc_random}{Random effects formula for the missingness indicators model for the costs}

\item{pe_random}{Number of random effects for the effectiveness model}

\item{pc_random}{Number of random effects for the cost model}

\item{ze_random}{Number of random effects or the missingness indicators model for the effectiveness}

\item{zc_random}{Number of random effects or the missingness indicators model for the costs}
}
\description{
This function modifies default hyper prior parameter values in the type of selection model selected according 
to the type of missingness mechanism and distributions for the outcomes assumed.
}
\examples{
#Internal function only
#no examples
#
#
}
\keyword{distributions}
\keyword{longitudinal}
\keyword{models}
\keyword{priors}
