% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_pattern.R
\name{prior_pattern}
\alias{prior_pattern}
\title{An internal function to change the hyperprior parameters in the pattern mixture model provided by the user depending on the type of
missingness mechanism and outcome distributions assumed}
\usage{
prior_pattern(type, dist_e, dist_c, model_txt_info, model_string_jags)
}
\arguments{
\item{type}{Type of missingness mechanism assumed. Choices are Missing At Random (MAR), Missing Not At Random for the effects (MNAR_eff),
Missing Not At Random for the costs (MNAR_cost), and Missing Not At Random for both (MNAR). For a complete list of all available hyper parameters 
and types of models see the manual.}

\item{dist_e}{distribution assumed for the effects. Current available choices are: Normal ('norm'), Beta ('beta'), Gamma ('gamma'), Exponential ('exp'),
Weibull ('weib'), Logistic ('logis'), Poisson ('pois'), Negative Binomial ('negbin') or Bernoulli ('bern')}

\item{dist_c}{Distribution assumed for the costs. Current available choices are: Normal ('norm'), Gamma ('gamma') or LogNormal ('lnorm')}

\item{model_txt_info}{list containing model specification information used to write the txt file of the JAGS model.}

\item{model_string_jags}{text file of the model.}
}
\description{
This function modifies default hyper prior parameter values in the type of pattern mixture model selected according 
to the type of missingness mechanism and distributions for the outcomes assumed.
}
\examples{
#Internal function only
#no examples
#
#
}
\keyword{Pattern}
\keyword{distributions}
\keyword{mixture}
\keyword{models}
\keyword{priors}
