% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{An internal function to read and re-arrange the data in different ways}
\usage{
read_data(data)
}
\arguments{
\item{data}{a data frame with rows representing the individuals and with columns that must included 
the following elements: a column named 'e' containing the effectiveness data for the individuals, a column named 'c'
containing the cost data for the individuals, a column 't' representing the tretment arm allocation for each individual 
(only two arms are supported with 1 being the control and 2 being the new intervention). Additional elements that could
be provided are: a number of columns named 'X1,X2,...' each containing different covariate data for the individuals. 
Covariate data must be fully observed and if categorical they must be provided as factor variables.}
}
\description{
This function imports the data and outputs all types of information used to run the model.
}
\examples{
\dontrun{
#create a data set which respects the requirements specified in "data" (see Arguments)
N1<-150
N2<-100
m_eff1<-m_cost1<-rbinom(N1,1,0.25)
m_eff2<-m_cost2<-rbinom(N2,1,0.25)
m_cost1<-m_cost1<-rbinom(N1,1,0.25)
m_cost2<-m_cost2<-rbinom(N2,1,0.25)
eff1<-rnorm(N1,0.5,0.5)
eff2<-rnorm(N2,0.5,0.5)
cost1<-rnorm(N1,90,20)
cost2<-rnorm(N2,90,20)
#introduce missingness
eff1[m_eff1==1]<-NA
eff2[m_eff2==1]<-NA
cost1[m_cost1==1]<-NA
cost2[m_cost2==1]<-NA
#arrange data frame
e<-c(eff1,eff2)
c<-c(cost1,cost2)
m_eff<-c(m_eff1,m_eff2)
m_cost<-c(m_cost1,m_cost2)
t<-c(t1,t2)
data<-data.frame(e,c,t)
#run the function
date_rearranged<-read_data(data=data)
}
#
#
}
\keyword{data}
\keyword{read}
