% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bugs.R
\name{run_bugs}
\alias{run_bugs}
\title{An internal function to execute a BUGS model and get posterior results}
\usage{
run_bugs(type, dist_e, dist_c, forward, inits)
}
\arguments{
\item{type}{Type of missingness mechanism assumed. Choices are Missing Completely At Random (MCAR),
Missing At Random (MAR), Missing Not At Random (MNAR). For 'MNAR' alternative versions are available 
depending on whether the mechanism for only one variable is condiered, that is for the effects (MNAR_eff)
or the costs (MNAR_cost), or also covariates are included either for the effects (MNAR_eff_cov),
the costs (MNAR_cost), or both (MNAR_cov).}

\item{dist_e}{effect data matrix where rows represent individuals and columns represent the arms (only 2 supported)}

\item{dist_c}{cost data where rows represent individuals and columns represent the arms (only 2 supprted)}

\item{forward}{Logical. If \code{forward} is \code{TRUE}, the model is run in forward sampling mode
without providing any data, else if \code{forward} is \code{FALSE} standard sampling mode is selected.}

\item{inits}{a list with elements equal to the number of chains selected; each element of the list is itself a list of starting values for the BUGS model, 
or a function creating (possibly random) initial values. If inits is NULL, BUGS will generate initial values for parameters}
}
\description{
This function fits a BUGS using the \code{\link[R2OpenBUGS]{bugs}} funciton and obtain posterior inferences.
}
\examples{
#Internal function only
#No examples
#
#
}
\keyword{BUGS}
\keyword{Bayesian}
\keyword{model}
