% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hurdle.R
\name{run_hurdle}
\alias{run_hurdle}
\title{An internal function to execute a JAGS hurdle model and get posterior results}
\usage{
run_hurdle(data_model, type, dist_e, dist_c, model_info)
}
\arguments{
\item{data_model}{list containing the data for the model to be passed to JAGS.}

\item{type}{Type of structural value mechanism assumed. Choices are Structural Completely At Random (SCAR), Structural At Random (SAR),}

\item{dist_e}{distribution assumed for the effects. Current available choices are: Normal ('norm'), Beta ('beta'), Gamma ('gamma'), Exponential ('exp'),
Weibull ('weib'), Logistic ('logis'), Poisson ('pois'), Negative Binomial ('negbin') or Bernoulli ('bern').}

\item{dist_c}{Distribution assumed for the costs. Current available choices are: Normal ('norm'), Gamma ('gamma') or LogNormal ('lnorm').}

\item{model_info}{list containing model and MCMC information to be passed to JAGS.}
}
\description{
This function fits a JAGS using the \code{\link[R2jags]{jags}} function and obtain posterior inferences.
}
\examples{
#Internal function only
#No examples
#
#
}
\keyword{Bayesian}
\keyword{JAGS}
\keyword{hurdle}
\keyword{models}
