% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lmdm.R
\name{run_lmdm}
\alias{run_lmdm}
\title{An internal function to execute a JAGS longitudinal missing data model and get posterior results}
\usage{
run_lmdm(data_model, type, dist_e, dist_c, model_info)
}
\arguments{
\item{data_model}{list containing the data for the model to be passed to JAGS.}

\item{type}{Type of missingness mechanism assumed. Choices are Missing At Random (MAR), Missing Not At Random for the effects (MNAR_eff),
Missing Not At Random for the costs (MNAR_cost), and Missing Not At Random for both (MNAR).}

\item{dist_e}{distribution assumed for the effects. Current available choices are: Normal ('norm'), Beta ('beta'), Gamma ('gamma'), Exponential ('exp'),
Weibull ('weib'), Logistic ('logis'), Poisson ('pois'), Negative Binomial ('negbin') or Bernoulli ('bern').}

\item{dist_c}{Distribution assumed for the costs. Current available choices are: Normal ('norm'), Gamma ('gamma') or LogNormal ('lnorm').}

\item{model_info}{list containing model and MCMC information to be passed to JAGS.}
}
\description{
This function fits a JAGS using the \code{\link[R2jags]{jags}} function and obtain posterior inferences.
}
\examples{
#Internal function only
#No examples
#
#
}
\keyword{Bayesian}
\keyword{JAGS}
\keyword{data}
\keyword{longitudinal}
\keyword{missing}
\keyword{models}
