% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Full Bayesian Models to handle missingness in Health Economic Evaluations}
\usage{
run_model(data, dist_e, dist_c, type, program = "JAGS", forward = FALSE,
  prob = c(0.05, 0.95), n.chains = 2, n.iter = 20000,
  n.burnin = floor(n.iter/2), inits = NULL, n.thin = 1,
  save_model = FALSE, prior = "default", ...)
}
\arguments{
\item{data}{A data frame with rows representing the individuals and with columns that must included 
the following elements: a column named 'e' containing the effectiveness data for the individuals, a column named 'c'
containing the cost data for the individuals, a column 't' representing the tretment arm allocation for each individual 
(only two arms are supported with 1 being the control and 2 being the new intervention). Additional elements that could
be provided are: a number of columns named 'X1,X2,...' each containing different covariate data for the individuals. 
Covariate data must be fully observed and if categorical they must be provided as factor variables.}

\item{dist_e}{Distribution assumed for the effects. Current available chocies are: Normal ("norm") and Beta ("beta").}

\item{dist_c}{Distribution assumed for the costs. Current available chocies are: Normal ("norm") and Gamma ("gamma").}

\item{type}{Type of missingness mechanism assumed. Choices are: Missing Completely At Random (MCAR),
Missing At Random (MAR), Missing Not At Random (MNAR). Different 'MNAR' alternative versions are available 
depending on whether the mechanism for only one or both outcomes is considered. Specifically it is possible
to select MNAR for the effect variables ('MNAR_eff'), for the cost variables ('MNAR_cost') or both ('MNAR').
If covariate data are provided in \code{data}, \code{run_model} automatically includes them in the model.}

\item{program}{type of software used to run the model. Current alternatives are 'OpenBUGS'('BUGS') and 'JAGS' ('JAGS').}

\item{forward}{Logical. If \code{forward} is \code{TRUE}, the model is run in forward sampling mode
without using any data, else if \code{forward} is \code{FALSE} standard sampling mode is selected.}

\item{prob}{A numeric vector of probabilities within the range (0,1), representing the upper and lower
CI sample quantiles to be calculated and returned for the imputed values.}

\item{n.chains}{Number of chains.}

\item{n.iter}{Number of iterations.}

\item{n.burnin}{Number of warmup iterations.}

\item{inits}{A list with elements equal to the number of chains selected; each element of the list is itself a list of starting values for the
\code{BUGS/JAGS} model, or a function creating (possibly random) initial values. If \code{inits} is \code{NULL}, \code{BUGS/JAGS}
 will generate initial values for all the model parameters.}

\item{n.thin}{Thinning interval.}

\item{save_model}{Logical. If \code{save_model} is \code{TRUE} a \code{txt} file containing the model code is printed
in the current working directory.}

\item{prior}{A list containing the hyperprior values provided by the user. Each element of this list must be a vector of length two
containing the user-provided hyperprior values and must be named with the name of the corresponding parameter. For example, the hyperprior
values for the mean effect parameter, when a normal distribution is assumed, can be provided using \code{prior=list('mu.prior.e'=c(0,1))}.
For more information about how to provide prior hypervalues for each model parameter see details. If \code{prior} is 'default' the default values will be used.}

\item{...}{additional input parameters. Examples are \code{ind} (logical), \code{stand} (logical) and \code{transf}. When \code{ind} is \code{FALSE} 
a joint model is run (only for bivariate normal), while when \code{stand} is \code{TRUE} outcomes are scaled so to have mean \code{0} and
standard deviation \code{0.5} (only for bivariate normal). If \code{transf} is set to 'logit', mean effect variables are modelled on the log-odds scale 
(only for beta distribution), while if 'log' is chosen, mean cost variables are modelled on the log scale (only for gamma distribution). If a beta-gamma
model is selected both transformations can be carried out by setting \code{transf} equal to a vector that contains both transformation string names.
Other additional arguments contained in the function \code{\link[BCEA]{bcea}} can be provided.}
}
\value{
An object of the class "missingHE" containing the following elements
\describe{
  \item{data_set}{A list containing the original data set provided in \code{data} (see Arguments), the number of observed and missing individuals 
  and the total number of individuals by treatment arm}
  \item{model_output}{A list containing the output of a \code{BUGS/JAGS} model generated from the functions \code{\link[R2jags]{jags}} 
  or \code{\link[R2OpenBUGS]{bugs}}}
  \item{mean_effects}{A matrix with \code{nsim} rows and \code{2} columns containing the posterior samples for the mean effect parameters in the
  treatment arms}
  \item{mean_costs}{A matrix with \code{nsim} rows and \code{2} columns containing the posterior samples for the mean cost parameters in the
  treatment arms}
  \item{sd_effects}{A matrix with \code{nsim} rows and \code{2} columns containing the posterior samples for the standard 
  deviation effect parameters in the treatment arms}
  \item{sd_costs}{A matrix with \code{nsim} rows and \code{2} columns containing the posterior samples for the standard 
  deviation cost parameters in the treatment arms}
  \item{imputed}{A list containing the the posterior samples for the imputed individuals in each arm and for each outcome.
  The stored imputed values represent the mean and the upper and lower quantiles of the posterior distribution for the
  missing individuals according to the values defined in \code{prob} (see Arguments)}
  \item{type}{A character variable that indicate which type of program has been used to run the model, either \code{JAGS} or \code{BUGS}}
}
}
\description{
Full Bayesian cost-effectiveness models to handle missing data in the outcomes under different missing data 
mechanism assumptions and using a variatey of alternative parametric distributions for the effect and cost variables and 
using a selection model to specify the missingness mechanism. The analysis is performed using the \code{BUGS} language, 
which is implemented either in \code{JAGS} or \code{OpenBUGS} using the functions \code{\link[R2jags]{jags}} and 
\code{\link[R2OpenBUGS]{bugs}}, respectively. The output is stored in an object of class "missingHE".
}
\details{
Depending on the distributional assumptions specified for the outcome variables in the arguments \code{dist_e} and
\code{dist_c} (Model of Analysis, MoA) and the type of missingness mechanism assumed in the argument \code{type} (Model of Missingness, MoM)
, different types of models are built and run in the background by \code{run_model}.
Specifically, the MoM module is handled using a Selection Model (SM) appraoch, which can be summarised with a simple example.
Consider a data set comprising a partially observed response variable \eqn{y}, the corresponding missing data indicator vector \eqn{m}
, and a fully-observed covariate \eqn{x}. Under the SM approach, the joint distribution \eqn{p(y,m)} is factored as the product of 
the marginal distribution \eqn{p(y)} and the conditional distribution \eqn{p(m|y)}.
\deqn{p(y,m|x,\theta^MoA,\theta^MoM)=p(y|x,\theta^MoA)p(m|y,x,\theta^MoM)}
where, \eqn{\theta^MoA},\eqn{\theta^MoM} are respectively the main parameters of interest (in the MoA) 
and the parameters associated with the missing data mechanism (in the MoM).
While the MoA specification depends on the assumed distributional for for \eqn{y}, the MoM can have a more generalised structure.
\code{run_model} assigns a Bernoulli probability distribution to the missing data indicator \eqn{m ~ Bernoulli(\pi)}, 
where \eqn{\pi} is the parameter specifying the probability of \eqn{y} being missing. Such probability is modelled using
a logistic regression transformation. More specifically:
\deqn{logit(\pi)=\gamma_0+\gamma_1 x+\delta y}
where
\itemize{
\item \eqn{\gamma_0} represents the baseline probability of missingness in \eqn{y} that does not depend on any variable. When \eqn{\gamma_1=\delta=0}, the model assumes MCAR.
\item \eqn{\gamma_1} represents the impact on the probability of missingness in \eqn{y} of the fully observed covariate \eqn{x}. When \eqn{\delta=0}, the model assumes MAR.
\item \eqn{\delta} represents the impact on the probability of missingness in \eqn{y} of the possibly unobserved values in \eqn{y} (MNAR).
}

It is important to keep in mind both the model structure of the MoA and MoM when providing user-defined hyperprior values for the parameters indexing such components
using the argument \code{prior} in the function \code{run_model}. Specifically, for the MoA module the default prior distributions that can be overwritten by the user
are the following:
\itemize{
\item Normal. Priors can be supplied for the location (mean) and auxiliary (log-sd) parameter with default values: \eqn{\mu ~ Normal(0,0.00001), \alpha~Uniform(-5,10)}
\item Beta.  Priors can be supplied for the location (mean) and auxiliary (sd) parameter with default values: \eqn{\mu ~ Uniform(0,1), \alpha~Uniform(0,\sqrt\mu(1-\mu))}
\item Gamma.  Priors can be supplied for the location (mean) and auxiliary (sd) parameter with default values: \eqn{\mu ~ Uniform(0,10000), \alpha~Cauchy(0,0.16)T(0,)}
}
When covariate data are included in the model with a linear regression to the mean of the outcome variable, then instead of specifying priors on \eqn{\mu} 
it is necessary to define the priors on the coefficient parameters of the linear regression such as
\deqn{\mu=\beta_0+\sum\beta_j X_j}
where \eqn{\beta_0} is the marginal mean to which the default prior is the same as that specified on \eqn{\mu} if no covariate is included, while
\eqn{\beta_j} are the covariate coefficients whose defualt prior is the same foe all j: \eqn{\beta_j~Normal(0,0.000001)}. 
For the MoM parameters the default prior distributions assumed are the following:
\itemize{
\item \eqn{gamma_0~Logisitc(0,1)}
\item \eqn{gamma_j~Normal(0,1)}
\item \eqn{delta~Normal(0,1)}
}
When user-defined hyperprior values are supplied via the argument \code{prior} in the function \code{run_model}, the list elements inside such an object (see Arguments)
must be vectors of length 2 containing the user-provided hyperprior values and must take specific names according to the parameters whose priors the user wants to modify. 
Specifically, the names accepted by \strong{missingHE} are the following:
\itemize{
\item location parameters in the MoA: "mean.prior.e"(effects) and/or "mean.prior.c"(costs)
\item auxiliary parameters in the MoA: "alpha.prior.e"(effects) and/or "alpha.prior.c"(costs)
\item marginal mean parameter (if covariate data provided) in the MoA: "beta0.prior.e"(effects) and/or "beta0.prior.c"(costs)
\item covariate parameters (if covariate data provided) in the MoA: "beta.prior.e"(effects) and/or "beta.prior.c"(costs)
\item baseline parameter in the MoM: "gamma0.prior.e"(effects) and/or "gamma0.prior.c"(costs)
\item covariate parameters in the MoM (if covariate data provided) in the MoM: "gamma.prior.e"(effects) and/or "gamma.prior.c"(costs)
\item mnar parameter in the MoM: "delta.prior.e"(effects) and/or "delta.prior.c"(costs)
} 
To notice that the marginal mean and covariate parameters can be provided only if covariate data are inlcuded in the model, with the
marginal mean parameter being a substitute of the location parameter under such model framework.
}
\examples{
#Simple example to simulate and analyse a data set
#Define the number of individuals fer treatment arm
N1<-150
N2<-100
#Create the misisngness indicators totally random (MCAR mechanism)
m_eff1<-m_cost1<-rbinom(N1,1,0.25)
m_eff2<-m_cost2<-rbinom(N2,1,0.25)
#Simulate data from normal distributions for both arms
eff1<-rnorm(N1,0.5,0.5)
eff2<-rnorm(N2,0.5,0.5)
cost1<-rnorm(N1,90,20)
cost2<-rnorm(N2,90,20)
#Set value missing if indicator is 1
eff1[m_eff1==1]<-NA
eff2[m_eff2==1]<-NA
cost1[m_cost1==1]<-NA
cost2[m_cost2==1]<-NA
#Create treatment arm indicators
t1<-rep(1,length(eff1))
t2<-rep(2,length(eff2))
#Combine variables and define a data set
e<-c(eff1,eff2)
c<-c(cost1,cost2)
m_eff<-c(m_eff1,m_eff2)
m_cost<-c(m_cost1,m_cost2)
t<-c(t1,t2)
data<-data.frame(e,c,t)

#Run the model using run_model with JAGS assuming a MCAR missingness mechanism
x<-run_model(data=data,dist_e="norm",dist_c="norm",type="MCAR",program="JAGS")
#
#
#Assess model convergence using graphical tools
#Produce histograms of the posterior samples for the mean effect
#parameters in the two treatment arms. 
diagnostic_checks(x,class = "histogram",parm = "mu.e")
#
#Compare observed outcome data with imputations from the model
# (posteiror means and credible intervals)
plot(x,class="scatter",outcome="all")
#
#Summarise the CEA information from model results
summary(x)
#
#
}
\references{
Daniels, MJ. Hogan, JW. \emph{Missing Data in Longitudinal Studies: strategies for Bayesian modelling and sensitivity analysis}, CRC/Chapman Hall.

Baio, G.(2012). \emph{Bayesian Methods in Health Economcis}. CRC/Chapman Hall, London.

Gelman, A. Carlin, JB., Stern, HS. Rubin, DB.(2003). \emph{Bayesian Data Analysis, 2nd edition}, CRC Press.

Plummer, M. \emph{JAGS: A program for analysis of Bayesian graphical models using Gibbs sampling.} (2003).
}
\seealso{
\code{\link[R2OpenBUGS]{bugs}}, \code{\link[R2jags]{jags}}, \code{\link[BCEA]{bcea}}
}
\author{
Andrea Gabrio
}
\keyword{CEA}
\keyword{JAGS}
\keyword{Missingness}
\keyword{OpenBUGS}
