% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_model.R
\name{write_model}
\alias{write_model}
\title{Alternatives vary depending on the type of distribution assumed for the effect and cost variables, choice of natural-scaled or scaled outcome
This function selects which type of model to execute.}
\usage{
write_model(type, dist_e, dist_c, program, forward)
}
\arguments{
\item{type}{Type of missingness mechanism assumed. Choices are Missing Completely At Random (MCAR),
Missing At Random (MAR), Missing Not At Random (MNAR). For 'MNAR' alternative versions are available 
depending on whether the mechanism for only one variable is condiered, that is for the effects (MNAR_eff)
or the costs (MNAR_cost), or also covariates are included either for the effects (MNAR_eff_cov),
the costs (MNAR_cost), or both (MNAR_cov).}

\item{dist_e}{distribution assumed for the effects. Current available chocies are: Normal.}

\item{dist_c}{distribution assumed for the costs. Current available chocies are: Normal.}

\item{program}{type of software used to run the model. Current alternatives are 'OpenBUGS' and 'JAGS'.}

\item{forward}{Logical. If \code{forward} is \code{TRUE}, the model is run in forward sampling mode
without providing any data, else if \code{forward} is \code{FALSE} standard sampling mode is selected.}
}
\description{
Alternatives vary depending on the type of distribution assumed for the effect and cost variables, choice of natural-scaled or scaled outcome
This function selects which type of model to execute.
}
\examples{
#Internal function only
#No examples
#
#
}
\keyword{BUGS}
\keyword{JAGS}
\keyword{models}
