% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_selection.R
\name{write_selection}
\alias{write_selection}
\title{An internal function to select which type of selection model to execute. 
Alternatives vary depending on the type of distribution assumed for the effect and cost variables,
type of missingness mechanism assumed and independence or joint modelling
This function selects which type of model to execute.}
\usage{
write_selection(type, dist_e, dist_c)
}
\arguments{
\item{type}{Type of missingness mechanism assumed. Choices are Missing At Random (MAR), Missing Not At Random for the effects (MNAR_eff),
Missing Not At Random for the costs (MNAR_cost), and Missing Not At Random for both (MNAR)}

\item{dist_e}{distribution assumed for the effects. Current available chocies are: Normal ('norm') or Beta ('beta').}

\item{dist_c}{distribution assumed for the costs. Current available chocies are: Normal ('norm') or Gamma ('gamma').}
}
\description{
An internal function to select which type of selection model to execute. 
Alternatives vary depending on the type of distribution assumed for the effect and cost variables,
type of missingness mechanism assumed and independence or joint modelling
This function selects which type of model to execute.
}
\examples{
#Internal function only
#No examples
#
#
}
\keyword{JAGS}
\keyword{Selection}
\keyword{models}
