% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.missingHE.R
\name{plot.missingHE}
\alias{plot.missingHE}
\title{Missing Data Imputation Plot}
\usage{
\method{plot}{missingHE}(x, prob = c(0.05, 0.95), class = "scatter",
  outcome = "all", theme = "base", ...)
}
\arguments{
\item{x}{A \code{missingHE} object containing the results of the Bayesian modelling and the economic evaluation}

\item{prob}{A numeric vector of probabilities within range [0, 1], representing the upper and lower
CI sample quantiles to be calculated and returned for the imputed values.}

\item{class}{Type of plot comparing the observed and imputed outcome data. Available choices are
'histogram' and 'scatter' for a histogram or a scatter plot of the observed and imputed outcome data, respectively.}

\item{outcome}{For which outcome type the plots should be displayed. Options are: 'all' (default) shows the plots
 for both treatment arms and types of outcome variables. 'effects' or 'costs' show the plots for the corresponding
 outcome variables in both arms, while 'arm1' or 'arm2' show the plots by the selected treatment arm. To select
 the plots for a specific type of outcome in a specific treatment arm the options that can be used are 'effects_arm1',
 'effects_arm2', 'costs_arm1' or 'costs_arm2'.
Treatment or outcome specific results can also be shown. For further details see the manual.}

\item{theme}{Type of ggplot theme among some pre-defined themes, mostly taken from the package \strong{ggthemes}. For a full list of available themes see details.}

\item{...}{additional parameters that can be provided to manage the output of \code{plot.missingHE}.}
}
\value{
A \code{ggplot} object containing the plots specified in the argument \code{class}.
}
\description{
Produces a plot of the observed and imputed values (with credible intervals) for the effect and cost outcomes from a simulation
Cost-Effectiveness Analysis model with two treatment arms that is run using the function \code{\link{run_model}}.
Graphics layout and management is obtained from the functions contained in the package \code{ggplot2} and \code{ggthemes}.
}
\details{
The funciton produces a plot of the observed and imputed effect and cost data in a two-arm based 
Cost-Effectiveness Analysis simulation model run with the function \code{\link{run_model}}. The purpose of this graph
is to visually compare the outcome values for the fully-observed individuals with those imputed by the model for the missing individuals.
According to the missing data mechanism assumption defined in \code{run_model}, it is possible to observe how
imputations may vary with respect to the observed cases. For the scatter plot, imputed values are also associated with
the credible intervals specified in the argument \code{prob}.
The argument \code{theme} allows to customise the graphical aspect of the plots generated by \code{plot.missingHE} and
allows to choose among a set of possible pre-defined themes taken form the package \strong{ggtheme}. Those available can
be indicated using the following character names: "base","calc","economist","excel","few","538","gdocs","hc","par","pander","solarized","stata","tufte","wsj".
}
\examples{
#For examples see the function run_model
#
#
}
\references{
Daniels, MJ. Hogan, JW. \emph{Missing Data in Longitudinal Studies: strategies for Bayesian modelling and sensitivity analysis}, CRC/Chapman Hall.

Molenberghs, G. Fitzmaurice, G. Kenward, MG. Tsiatis, A. Verbeke, G. \emph{Handbook of Missing Data Methodology}, CRC/Chapman Hall.
}
\seealso{
\code{\link{run_model}} \code{\link{diagnostic_checks}}
}
\author{
Andrea Gabrio
}
\keyword{data}
\keyword{missing}
\keyword{plot}
