% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.missingHE.R
\name{print.missingHE}
\alias{print.missingHE}
\title{Print a summary of the \code{JAGS} model fitted by \code{\link{selection}} or \code{\link{hurdle}}}
\usage{
\method{print}{missingHE}(x, value.mis = FALSE, ...)
}
\arguments{
\item{x}{A \code{missingHE} object containing the results of the Bayesian model run using the function \code{\link{selection}} or \code{\link{hurdle}}.}

\item{value.mis}{Logical. If \code{value.mis} is \code{TRUE}, the model results displayed contain also the imputed values,
else if \code{value.mis} is \code{FALSE} the missing values are hidden.}

\item{...}{additional arguments affecting the printed output produced. For example: \code{digits=} number of 
significant digits to be shown in the printed table (default=3). Not available if \code{value.mis=}TRUE.}
}
\description{
Prints the summary table for the model fitted, with the estimate of the parameters and/or missing values.
}
\examples{
 
#For examples see the function selection or hurdle 
# 
# 
}
\author{
Andrea Gabrio
}
\keyword{JAGS}
\keyword{data}
\keyword{missing}
\keyword{print}
