% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_hurdle.R
\name{prior_hurdle}
\alias{prior_hurdle}
\title{An internal function to change the hyperprior parameters in the hurdle model provided by the user depending on the type of
structural value mechanism and outcome distributions assumed}
\usage{
prior_hurdle(type, dist_e, dist_c)
}
\arguments{
\item{type}{Type of structural value mechanism assumed. Choices are Structural Completely At Random (SCAR),
and Structural At Random (MNAR). For a complete list of all available hyper parameters 
and types of models see the manual.}

\item{dist_e}{distribution assumed for the effects. Current available chocies are: Normal ('norm') or Beta ('beta').}

\item{dist_c}{distribution assumed for the costs. Current available chocies are: Normal ('norm') or Gamma ('gamma').}
}
\description{
This function modifies default hyper prior parameter values in the type of hurdle model selected according 
to the type of structural value mechanism and distributions for the outcomes assumed.
}
\examples{
#Internal function only
#no examples
#
#
}
\keyword{distributions}
\keyword{hurdle}
\keyword{models}
\keyword{priors}
