% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.missingHE.R
\name{summary.missingHE}
\alias{summary.missingHE}
\title{Summary method for objects in the class \code{missingHE}}
\usage{
\method{summary}{missingHE}(object, ...)
}
\arguments{
\item{object}{A \code{missingHE} object containing the results of the Bayesian modelling and the economic evaluation}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
Prints a summary table with some information on the health economic simulation model run with the 
function \code{run_model}, and synthetic information on the main parameters of interests.
}
\description{
Produces a table printout with some summary results of the health economic evaluation simulation model
 run using the function \code{\link{run_model}}.
}
\examples{
#For examples see the function run_model
#
#
}
\references{
Baio, G.(2012). \emph{Bayesian Methods in Health Economcis}. CRC/Chapman Hall, London.
}
\seealso{
\code{\link{run_model}} \code{\link{diagnostic_checks}} \code{\link{plot.missingHE}}
}
\author{
Andrea Gabrio
}
