% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.missingHE.R
\name{summary.missingHE}
\alias{summary.missingHE}
\title{Summary method for objects in the class \code{missingHE}}
\usage{
\method{summary}{missingHE}(object, ...)
}
\arguments{
\item{object}{A \code{missingHE} object containing the results of the Bayesian modelling and the economic evaluation}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
Prints a table with some information on the health economic model based on the assumption
selected for the missingness ('MAR'or 'MNAR') or structural values ('SCAR' or 'SAR') mechanism using the function \code{selection} or \code{hurdle}, respectively. 
Summary information on the main parameters of interests is provided.
}
\description{
Produces a table printout with some summary results of the health economic evaluation probabilistic model
run using the function \code{\link{selection}} or \code{\link{hurdle}}.
}
\examples{
#For examples see the function selection or hurdle
#
#
}
\references{
Baio, G.(2012). \emph{Bayesian Methods in Health Economcis}. CRC/Chapman Hall, London.
}
\seealso{
\code{\link{selection}} \code{\link{hurdle}} \code{\link{diagnostic}} \code{\link{plot.missingHE}}
}
\author{
Andrea Gabrio
}
