% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hurdle.R
\name{write_hurdle}
\alias{write_hurdle}
\title{An internal function to select which type of hurdle model to execute. 
Alternatives vary depending on the type of distribution assumed for the effect and cost variables,
type of strcutural value mechanism assumed and independence or joint modelling
This function selects which type of model to execute.}
\usage{
write_hurdle(type, dist_e, dist_c, se = se, sc = sc)
}
\arguments{
\item{type}{Type of structural value mechanism assumed. Choices are Structural Completely At Random (SCAR),
and Structural At Random (MNAR)}

\item{dist_e}{distribution assumed for the effects. Current available chocies are: Normal ('norm') or Beta ('beta').}

\item{dist_c}{distribution assumed for the costs. Current available chocies are: Normal ('norm') or Gamma ('gamma').}

\item{se}{Structural value to be found in the effect data. If set to \code{NULL}, 
no structural value is chosen and a standard model for the effects is run.}

\item{sc}{Structural value to be found in the cost data. If set to \code{NULL}, 
no structural value is chosen and a standard model for the costs is run.}
}
\description{
An internal function to select which type of hurdle model to execute. 
Alternatives vary depending on the type of distribution assumed for the effect and cost variables,
type of strcutural value mechanism assumed and independence or joint modelling
This function selects which type of model to execute.
}
\examples{
#Internal function only
#No examples
#
#
}
\keyword{JAGS}
\keyword{hurdle}
\keyword{models}
